+++ modules/admin/tests/admin/commands/CrudControllerTest.php
@@ -1,12 +1,12 @@
 <?php
 
-namespace luyatests\core\console\commands;
+namespace admintests\admin\commands;
 
 use Yii;
-use luyatests\LuyaConsoleTestCase;
-use luya\console\commands\CrudController;
+use luya\admin\commands\CrudController;
+use admintests\AdminTestCase;
 
-class CrudControllerTest extends LuyaConsoleTestCase
+class CrudControllerTest extends AdminTestCase
 {
     public function testFindModelFolderIsModelFolderAvailable()
     {
@@ -41,7 +41,7 @@ namespace file\namespace;
 /**
  * Test Model.
  * 
- * File has been created with `crud/create` command on LUYA version 1.0.0-dev. 
+ * File has been created with `crud/create` command on LUYA version 1.0.0-RC3. 
  */
 class TestModel extends \luya\admin\ngrest\base\Api
 {
@@ -62,7 +62,7 @@ namespace file\namespace;
 /**
  * Test Model.
  * 
- * File has been created with `crud/create` command on LUYA version 1.0.0-dev. 
+ * File has been created with `crud/create` command on LUYA version 1.0.0-RC3. 
  */
 class TestModel extends \luya\admin\ngrest\base\Controller
 {
@@ -86,7 +86,7 @@ use luya\admin\ngrest\base\NgRestModel;
 /**
  * Test Model.
  * 
- * File has been created with `crud/create` command on LUYA version 1.0.0-dev. 
+ * File has been created with `crud/create` command on LUYA version 1.0.0-RC3. 
  *
  * @property integer $id
  * @property string $firstname
@@ -256,7 +256,7 @@ use luya\admin\ngrest\base\NgRestModel;
 /**
  * Test Model.
  * 
- * File has been created with `crud/create` command on LUYA version 1.0.0-dev. 
+ * File has been created with `crud/create` command on LUYA version 1.0.0-RC3. 
  *
  * @property integer $id
  * @property string $name