+++ modules/admin/tests/data/modules/crudmodulefolder/admin/Module.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace admintests\data\modules\crudmodulefolder\admin;
+
+class Module extends \luya\base\Module
+{
+}